Issue #3 - Output 5 most recent tweets
The program successfully outputs the 5 most recent tweets
Every 10 minutes the program will check and display new tweets

The program will tell the difference by hashing both old and new tweets.
If the hash is the same, the user has not posted. If it has, the program
will investigate and pull new tweets.

BUG:
When the user deletes a post, one of the older posts appears.